feat(PoolStore): create a dynamic pool store
This commit adds in a new service `PoolStore` to solve the issues of
selectively hiding data from lists when "used" and brought back when
"freed".  It includes Karma unit tests.

The scenario to think about is a typeahead that must hide data selective
when it is used. This allows dynamic lists to be formed where the data
needs to be hidden after used.