feat(store): implement store.clear()
This commit implements and tests the store.clear() method, useful for
removing all data from a store without destroying the reference to the
array.  Also contains some minor fixes for the unit tests of both the
PoolStore and the Store.

Finally, the services depending on store have been updated so as not to
break with the new store improvements.