refactor(invoice): update invoice to standards
This commit undertakes a massive refactor of the Patient Invoice module.
It tries to reduce complexity of the module by moving logic into
composable classes using services.

Here is the service hierarchy so far (subject to future improvements):
 1. `InvoiceService`
  1. `InvoiceItemService`
   1. `PatientInvoiceItemService`

Much of the logic has been moved out of the PatientInvoiceController and
into the InvoiceService.  The InvoiceService is mainly concerned with
the overall form (description, services, etc) and gathering the
appropriate metadata for the form.

The `InvoiceItemService` is concerned with the grid - totalling items
and dealing with the cache.  Note: the cache is a TODO.

The `PatientInvoiceItemService` is a simple class that composes each row
of the grid and allows for row-level validation.

This new service structure takes advantage of the PoolStoreService to
manage inventory items in the grid without having to manually manipulate
two stores.