chore(invoice): ensure grid totals resize properly
This commit adds an event to listen to the navigation change to ensure
that the grid totals row is properly (re)drawn when the tree is toggled.
However, a major caveat is that it requires a HARDCODED size of the tree
(260px).  This should be discussed and revisited at some point.