feat(invoice): implement caching
This commit implements caching in the PatientInvoiceFormService.  The
cache is currently never cleared, as it is not clear when it should be -
however, the hooks are implemented and work as expected!

This commit cleans up some of the HTML on the patient invoice page and
importantly finished documenting the Invoice.js Service.  A popover has
been implemented on the "Distribution Type" label to explain what it
actually does.

There are a few todos:
 1. Remove the InvoiceItems service
 2. Caching is blocked by a lack of API exposed by bhFindPatient.  This
 has been documented in #461.
 3. Clean up PatientInvoiceController.
 4. Rename Invoice.js to PatientInvoiceFormService