refactor(client): rm unused/old client modules (#496)
This commit removes or refactors a huge portion of the client.  It
mainly removes services which are no longer used, such as `connect.js`,
`compile.js` and others.  You can see the full list in the commit log.

In order to completely remove the `FormStateFactory`, it was necessary
to rewrite much of the Enterprise, Project, and Cashbox management
modules.  These have been partially brought up to 2.x standards with
regards to notifications and state management.  Cashboxes could still
use a rewrite that uses `ui-router` instead of a single page `vm.view`
hack.  The others are simple enough that is acceptable for now.

The tests have been updated to make sure they check for notifications
and ensure that notifications dismiss correctly.

Other notes:
 - All of `reports_proposed/` on the client has been removed.
 - Many of the old directives (like `ngBlur.js`) have been removed.  Any
 that broke the application were subsequently re-introduced.
 - The `/creditors` page is now `suppliers`. Closes #229.

This refactor should put us in a good position to tackle #495 next.