Re: AccountGrid service; guarantee account order
This commit re-factors the account grid service to remove the dependency
on accounts being sorted correctly. By always formatting/ using a single
store for accounts it does not matter if data is passed in from a cache
or the server. Simply sorting the account list also simplifies a lot of
the display logic.

* Re-factor AccountGrid Service
* Clean up Account Management controller, removing unneeded dependencies
* Moved grid templates into individual files