chore(services): improve & test API service
This commit improves the prototype API service with a cleaner API for
creating new services that inherit from it.  The new syntax is this:
```js
var service = new PrototypeApiService('url/');
// you can now call service.create(), service.update(), etc.
```

Additionally, the service has a new method `search()` to standardize the
way we send search queries.

Finally, the Prototype API service is now well tested with karma unit
tests.

Closes #421.