feat(purchase): purchase client creates orders
This commit hooks up the client services with the database to ensure
that purchase orders can be posted to the server.  In order to
accomplish this, I changed the schema slightly to account for the
new `payment_method` field.  This field might need to be an ID instead
of a text field, but this works for a first implementation until we
decide upon the value of the method.

Next, a purchase receipt should be implemented followed by end to end
tests.