fix(filters): ensure that filter clear is visible (#540)
This commit ensures that the bhFilterBar never covers the "clear all
filters" button to ensure that the end to end test never fail and the
user is always able to click the button.

Unfortunately, this also makes it so that `flex-content` overflows
vertically on the grid, which is not the best user experience.  We
should investigate this CSS issue soon to make sure that the UX is still
preserved.