Introduce angular-ui-select; update account create
This commit introduces the angular-ui-select library and includes any
required references to javascript/ CSS. An additional directive is
required to correctly focus the directive until this is addressed in the
library (or in angular-animate, the conflicting library).

angular-ui-select has been implemented in the Account Management create
state allowing users to select from a list of accounts.