refactor(invoices): leverage bhFiltersApplied
This commit refactors the modal form of the invoices registry to return
a simple, well-formatted JSON object.  The JSON object is then cached in
the invoice registry's local cache, and passed to bhFiltersApplied for
display.

It also removes the need for `$scope` in bhFiltersApplied.  Instead, I
have refactored the component to use the Angular 1.5 lifecycle hooks.
Specifically $onChanges() is called whenever the one-way data binding
changes.