test(invoice): refactor e2es for registry
This commit refactors the end-to-end tests for the registry, separating
out the search from the regular registry tests.  It also tries to go for
specific tests rather than every test possible to speed up e2e testing.

Testing in this manor revealed a bug with date processing - if only a
single date was provided, the database would respond with a 400 error.
This query has been updated so that we can limit by a single date at a
time.

Finally, the filter clearing as been tested.