feat(bhFindPatient): onRegisterApi callback
This commit introduces the onRegisterApi callback to the bhFindPatient
directive.  Any controller that needs access to the bhFindPatient's api
methods can request it via the `onRegisterApi` function.  For
demonstration, the Patient Invoice module does exactly this.

It also adds the initial unit test structure for karma testing the
component controller.  These unit tests will be rounded out in
subsequent commits.