refactor(journal): re-write grid column service
Closes #535.

This commit refactors the journal's column services into a single
service that can be shared across any UI grid.  It uses class instances
to namespace grids, while efficiently sharing methods between instances
via prototypes.

GridColumnService now expects the default visibility of columns to be
set on the column definitions in the gridOptions.  For example:
```js
$scope.gridOptions = {
  columnDefs : [{
    field : 'currency_id', displayName: 'Currency ID', visible: false
  }, {
    field : 'debit', displayName: 'Debit', visible: true
  }]
};
```

Note that ui-grid defaults to `visible:true` for all columns whose
visibility is not explicitly defined.

Additionally, the end-to-end tests have been rewritten to test the
following:
 1. Column selection checkboxes change the columns to the correct column
 names.
 2. Column selections are cached (survive browser refresh)
 3. Column config modal resets the default number of columns.

This PR does not completely address #376 yet.  Since there is some
non-standard service architecture, I am submitting it for review before
continuing on in this fashion.