fix(grid): transition journal to grid services
Closes #376.
Closes #586.

This commit completes the migration of specific posting journal services
to grid services.  It also removes the ability to paginate the posting
journal as this increases the complexity of the process several fold.

In addition to migrating the service, the GridFiltering service now
provides the option to toggle (inline) filtering on and off.  This could be
useful for power users wanting to filter the view quickly.