test(util): add AOP methods + test to util service
This commit adds aspect-oriented programming methods to the util service
for better readability of Grid Services.  It adds unit tests that show
functionality but do little to assure correctness.  These should be
expanded in the future as more use-cases for AOP become evident.

Most of the Grid Services use `onRegisterApi()` to bind the grid's API
to the service instance and perform specific actions before or after the
API becomes available.  This is the perfect use-case for aspect oriented
programming, so some naive methods have been implemented in the `util`
service library.

Additionally, this commit renames files without changing their contexts
for better clarity.