feat(journal): only edit editable transactions
This commit implements highlighting and edit mode for transaction groups
in the journal ui-grid.  It also ensures that only editable transactions
(those marked with _editing) are allowed to receive edits.

Finally, an editing flag is visible outside of the ui-grid to show that
there are currently transactions open for editing.