refactor(test): GridUtils for ui-grid e2e tests
This commit implements a helper library for ui-grid end-to-end tests.
The goal is to ensure uniform behavior between tests, increase ui-grid
test development, and ensure that dangling DOM nodes are not queried
after updates.  This is achieved by re-specifying the grid by ID each
time an action is to be taken on the grid.

Closes #614.