feat(db): automatically retry txns on deadlocks
This commit implements automatic retrying of transactions when a
transaction deadlock occurs.  This allows the server to handle up to 20
simultaneous patient registrations!  The integration tests have been
updated accordingly to stress the server.

The number of retries is currently 3, but is configurable via the
MAX_TRANSACTION_DEADLOCK_RESTARTS constant in the transaction library.

Closes #620.