refactor(invoice): optimize invoice POST data
Closes #579.

This commit refactors the patient invoice create API to only send back
the numeric billing service and subsidies IDs when creating new records.
Since all the information is already on the server, this was discarded
upon arrival anyway.  The client now only sends an array of IDs instead
of an array of objects.

Integration tests have been updated accordingly.