chore(ci): added fixtures directory, upload
This commit adds a fixtures directory to the test directory to store and
share fixtures between tests.  This allows us to reuse fixtures between
end to end and integration tests.

It also adds skeletal code for uploading screenshots to Amazon S3.  This
still needs some work - the Travis documentation might be out of date.