refactor(template): separate helpers and partials
This commit separates helpers and partials from the origin template.js
file into subdirectories of a template/ folder.  This directory can be
renamed as needed.  We now include partials for both the report
document head and report headers.

Additionally, the hard-coded styles have been moved into
bhima-bootstrap.less.

Next, most reports should be migrated to using this syntax for their
imports to reduce the amount of code we need to write to generate a new
report and numbers of variables we need to keep in our heads.