chore(report): use partials for condensed reports
This commit changes most reports to use partials, at least for the HTML
`<head>` tag.  It also cleans up the voucher reports to no longer need
to use `model.lang` to specify the language.

The final piece is to add report integration test to make sure that the
report will always be rendered.