fix(cash): rework transfer modal with ui-router
This commit fixes the cash transfer modal to work with the newer journal
vouchers interface.  The cash transfers now display the account names
and the currency is dependent on the account being selected.

The cash transfer modal can now be deep-linked via ui-router.  This
allows for a nice cleanup of states, but complicates the receipt modal a
little bit.  A receipt has not been implemented for cash transfers yet.