fix(bhCurrencySelect): do not select disabled vals
This commit fixes a bug that will select the provided currencyId even if
that currency is disabled.  It also improves performance by using
Angular's component hooks for initialization and watching changes.  The
disabled states are far more obvious using `text-muted`.

Finally, an HTML "title" attribute is used on disabled radio buttons to
make it abundantly clear why the radio button is disabled.