refactor(patient): use ui-router, ui-select
This commit refactors the patient routes to use `ui-router` for all
sub-routes.  These routes are all found in the module `bhima.routes` and
the file `patients/patients.routes.js`.

The patient registration page also leverages `ui-select` to select a
debtor group for the patient.  The end-to-end test has been updated
accordingly.