feat(cash): Posting Cash Payments
This feature allows the module to post cash payments with cash items.
To accomplish this, a new file called `functions.sql` was added to the
build scripts since `procedures.sql` was getting far to large.  This
file contains functions used in the procedures or other queries.

Posting cautions is not yet supported.