refactor(build): rm old/unused build modules
This commit removes old and unused build modules as recommended by the
comments in the gulpfile.  It should lighten the load in `npm install`
in the future.

It also introduces a PRODUCTION switch in the app.js file to make sure
the app runs in production mode and not in development mode by turning
off `$compileProvider` features.  Not all of these options are
available yet, but soon will be.  Note that this will need more testing
before launch.