refactor(tree): clearer UI + test hooks
This commit refactors the tree to include test hooks (`data-unit-key`)
and modifies the icons to have better contrast between folders and
files.

Bug fix: the navigation remembers the selected module and re-selects it
on `$stateChangeSuccess`.

Finally, test files have been added for (skipped) protractor tests to ensure
that the tree keeps working as expected.

Closes #44.