refactor(reports): prepare saved reports dir
This commit adds in a directory for saving reports, as well as adding a
the agedDebtors report page.  It adjusts some options in travis to try
and make the build more stable.

The ReportManager is now put on hold pending the installation in
our next site.  These advanced features will be worked on after the
initial installation occurs.