feat(entities): introduce entity_map
This commit creates a table that contains a mapping between entity uuids
and their human readable text.  All updates to this table are made via
triggers, leaving the controller code on both the server and the client
the same.

The entities are:
 1. Suppliers
 2. Employees
 3. Patients

NOTE that entities are both creditors and debtors.  Another table will
be made for documents (vouchers, invoices, payments, etc).

I've moved all triggers into triggers.sql.

Closes #753.

Closes #629.  Partially addresses #630.