fix(supplier): generate creditors from groups
This commit fixes a long-standing bug in the suppliers page.  The
suppliers previously used Creditors instead of Creditor Groups which
have no way to be created independently of the supplier.  The client
page has been refactored to choose creditor groups from a dropdown and
the server now generates a new creditor UUID when creating a supplier.

Closes #754.