refactor(creditNote): reverse invoices generically
This commit implements the generic reversals for the credit note route.
The generic reversal preserves the original transaction and simply flips
the debits and credits.

It also fixes a bug where templates were cached beyond their lifetime by
Angular.  For development purposes, templates are not cached.  In
production mode, the will be cached.

Closes #659.