feat(invoice): invoice caching on submit
This commit implements caching of patient invoices on submit.  When an
invoice is submitted, regardless of errors the form is cached in
AppCache.  Once the user successfully submits the form, the cache is
cleared.

The cache is only loaded once the user has chosen a patient.  The idea
is that a patient should not be cached, only the form data.

Finally, I've introduced an error flash on invalid grid rows.  To prompt
the error flash, one simply has to call `validate()` with the boolean
`true`.