Disable TTL on redis server (#859)
This commit sets the `disableTTL` flag on the connection request to the
redis session server.
As this flag has not been set before the server has been using the
default TTL for redis sessions of 1 day. If the server is running in
production session expiry is still used as default.