test(report): add generic report creation tests
This commit adds tests for creating a chart of accounts via the generic
report page.  It also adds a skeleton of what generic report page tests
should look like.

In the future, the generic report page should probably receive a
function to call instead of a JSON object for their creation form.  But
this can be worked out as more and more reports are added.