feat(journal): implement advanced filtering
This commit implements advanced filtering on the Posting Journal through
a modal form.  The module uses bhFiltersApplied to display the filters
currently applied to the posting journal at the top of the dash.

Filters are cached so that they survive page reloads.  All filtering is
done using the journal API from the server.

See #840.  Closes #793.