fix(errors): back button works on 403
This commit refactors the error pages to use the same folders and test
structure.  It moves some of the CSS to a central place and makes sure
that authentication works appropriately.

The most important change is that ALL states are sent to the client
along with their authorization status.  This is so that the app can tell
the difference between a 404 and a 403 effectively.

Closes #846.