fix(cash): alert when no invoices are selected
This commit introduces a check to verify that invoices are being paid
against when a user submits the cash form.  If there are no invoices, a
danger notification is emitted telling the user to add some invoices and
no HTTP request is sent.

Previous, this would be validated on the server - the client no longer
even needs to make the HTTP request.

Closes #772.