fix(vouchers): disable title accounts
This commit disables the title accounts on all voucher pages.  The title
accounts are still shown, but are disabled by checking the TITLE
bhConstant value.

It also does not allow a user to use the same account twice in a simple
voucher, as this would lead to an invalid transaction.

Closes #884.  Closes #896.