refactor(inventory): ui-route for create state
This commit breaks out the inventory routes into their own
inventory.routes.js file.  This file sets up the create state in a modal
for the inventory items to make it easier to navigate to.  Eventually,
it should implement a checkbox to add many items, such as the account
page.

It also improves the create form by giving it a bh-loading-button, and
cleaning up some of the validation.  Notify.handleError has now been
implemented, instead of the previous bug in error handling.