Fix inventory list (#929)
* fix(exchange): broke home page link

This commit fixes the broken update exchange rate link on the home page.
It now correctly points to the enterprise management page.

* fix(inventory): eliminate pagination

This commit removes pagination in the inventory list.

* fix(inventory): create/update error states

This commit ensures that the create/update modal properly handles error
state and reporting.  It also improves the UI slightly.

* fix(inventory): order inventory items on download

This commit changes the server to send back an ordered list of inventory
items, ordered by their code.

* refactor(inventory): ui-route for create state

This commit breaks out the inventory routes into their own
inventory.routes.js file.  This file sets up the create state in a modal
for the inventory items to make it easier to navigate to.  Eventually,
it should implement a checkbox to add many items, such as the account
page.

It also improves the create form by giving it a bh-loading-button, and
cleaning up some of the validation.  Notify.handleError has now been
implemented, instead of the previous bug in error handling.

* refactor(inventory): cosmetic improvements

This commit adds a headercrumb to the inventory modal, as well as
refactors the create/edit state handling within the modal controller.
Finally, it ensures that a notification is shown when the create/update
state successfully completes.

* fix(report): print-friendly tables

This commit adds rules to enable print-friendly tables in wkhtmltopdf.
It adds page break rules to indicate that pages should not break mid-row
in an overflow.

* refactor(inventory): refactor modal to use $state

This commit fixes some important bugs in the inventory update modal.  In
particular, the modal will wait to bind the inventory uuid in the UPDATE
state until it has all the type, unit, and group information.

It also completes the $state-based routing for the inventory modal.

* refactor(inventory): use ProtypeApiService

This commit migrates a lot of the dependent Inventory Services to use
PrototypeApiService.

* fix(Notify): errorHandler --> handleError

This commit changes all `Notify.errorHandler()` calls to
`Notify.handleError()`, throughout the entire application.