feat(bhAccountSelect): implement uiSelect account
This commit implements a generic account select for the app called
bhAccountSelect.  It uses a uiSelect, but introduces sane defaults.
This is tested on both the billing_services and Debtor Groups modules.

The defaults used are:
 1. debounce input by 200 milliseconds for better responsive behavior
 2. order accounts in hierarchical order
 3. disable title accounts

For performance reasons, this is the first component that uses one-way
bindings exclusively.  That means controllers will have to implement
callbacks to change the model data.

The usage is like this:
```html
<bh-account-select
  initial-value="$ctrl.model.account_id"
  on-update="$ctrl.updateAccount(account)"
  validation-trigger="Form.$submitted"
  required="true">
</bh-account-select>
```

The initial model value is the account id, and is set via the
`initial-value` parameter.  All updates are internal the component ...
to get the changed values, you must subscribe to the `on-update`
function callback.