fix(invoice): block no sales account
This commit ensures that the patient invoice page blocks invoices that
contain items missing sales account.  In production, this should only
occur if your database was improperly initialized.

Each item now is able to alert its own message as a tooltip when the
error state is hovered over.  Additionally, the grid itself displays a
large error when it detects items without a sales account.

Partially addresses #980.