(Feature) Add optional flag to use the combined ledger; debtor debt
This commit adds the `combinedLedger` flag to the aged debtor debt
report. This changes the 'source' of the debt query, using the
`general_ledger` by default and using the `combined_ledger` if the
flag is set.

It also updates the combined ledger to use the same column names as
the general ledger, querries can alias this after it is requested
if required.