Implement aggregate data flag for journal transactions
This commit adds the optional flag `aggregates` to the `/journal`
route. Setting the flag `aggregates=1` will now return a javascript
object with `journal` (rows) and `aggregate` data. Aggregate data will
provide the sum of debits and credits as well as the total number of
rows for every transaction in the journal query.

Aggregates have been tested to work for both standard requests and with
any of the filters applied.