Update transaction debits and credits; aggregates + header row
This commit introduces a basic pre processor for posting journal data,
this allows the header rows to show transaction level data even when
the journal is being filtered.

It also updates to using `uiGridGroupingConstants` aggregation types,
these aggregation methods are not altered when rows are expanded/
collapsed in the grid. This partially addresses #1041.

Each row now tracks information about the transaction as a whole.
This is used to include formation about the transaction in the row
header and also addresses #1041.

Closes #1041.
Closes #1045.