perf(settings): use translate directive
This commit modifies the tree and settings page to use the `translate`
directive instead of the translate filter.  This result in an average
decrease of `$watchers` by 50%.  The original page average was 648
`$watchers` with the tree fully expanded, but falls to 351 `$watchers`
by using the translate directive.

This commit determines the validity of #1074 and closes #1074.