fix(cash): payments registry filters work
This commit reimplements the payments registry filtering scheme in a way
that can be cached and improved upon later.  Previously, filters were
not able to be removed and frequent errors were thrown in the console.

It also refactors the routes so that all the payments routes are
together in the `cash` folder.  This is more in line with how the
invoices work.