fix(inventory): pass uuids in URL
This commit fixes the inventory server UPDATE route to conform to our
HTTP API standards.  It sends back a JSON object instead of an array,
allowing both the create() and update() routes to pass the uuid back to
the central list state and use the grid to scroll to the updated record.

This commit contains much performance refactoring.
  1. Makes the <select>s use track by in the ActionForm.
  2. Uses <translate> directive instead of filter
  3. Uses a loading indicator on the bhLoadingButton

A few of the issues in
https://github.com/IMA-WorldHealth/bhima-2.X/issues/738
have been addressed.