refactor(registries): grow shrink with filters (#49)
* fix(registries): grow/shrink with filter bar

This commit refactors the patient and invoice registries to respond to
the filter bar by growing and shrinking to take up as much space as
possible using `ngStyle`.  This allows the grid to display as many
records as possible in all instances.

* fix(registries): sort by number/date as needed

This commit ensures that the age and cost are identified as numeric
types and then sorted correctly.  It also adds the 'date' type to date
columns to ensure they sort nicely as well.

* test(invoices): add clear button tests

This commit adds a test to the invoice registry to clear the filters on
the invoice registry page and ensure all goes back to normal.  It also
fixes some minor translation issues (CASH.TRANSFER.SUCCESS) and a search
bug in the cash payments module.

* fix(cash): invoice IDs updated and sorted

This commit fixes the invoice lookup modal to properly sort by the date
DESC on the server and render the invoice identifiers in the new format.

* fix(enterprises): remove debounce from accounts

This commit removes the debounce from the ui-select for loss/gain
accounts.