fix(location:  unique key constraints (#57)
* fix(locations): rm unique keys on names

This commit removes unique keys on the name s of a villages, sectors,
and provinces.  The key is now made to take in both the village name and
the sector_uuid, and so on, so that no two named villages are in the
same sector.

Two integration tests have been added to validate this behavior.