refactor(inventory): action modal cleanup
This commit massively cleans up the action modal.  The modal allows
models to be updated independently for inventory groups, units, and types
and no longer wraps all in a promise.  The inventory services are
grouped under the top level inventory service.

The action form also uses util.filterFormElements to collect the data
from the form, and the (logically) equivalent Inventory.clean() function
has been removed.

The Inventory integration tests are better reflective of the behavior of
the module and no longer require all parts of information to work.