(fix) Ensure receipts have access to bootstrap classes
This commit uncomments the <link> tag to bootstrap CSS
in the header file shared by all larger form reports/ receipts.

In the future it would be good to remove this dependency on
bootstrap as very few of the classes are actually required or
used. A simple column library could be used to provide row/
column functionality.

It also adds CSS classes to ensure tables break over multiple
pages without colliding body/ header elements.