feature(report): support rendering to CSV
This commit implements the `csv` renderer using the `json-2-csv` module
on the server.  It can be accessed by simply requesting the
`renderer=csv` flag from the client.

The server must now specify a `csvKey` flag on all reports that will
support the csv renderer.  It looks like this:

```
function someReportRoute(req, res, next) {

  const options = _.clone(req.query);
  _.extend(options, { csvKey : 'accounts' });

  let report;

  try {
    report = new ReportManager(template, metadata, options);
  } catch (e) {
    return next(e);
  }

  Accounts.read()
    .then(accounts => report.render({ accounts }))
    .then(results => res.set(results.headers).send(results.report))
    .catch(next)
    .done();
}
```

The `csvKey` will set the object key that will be attempted to be
rendered to CSV.  This should be an array of JSON documents.

The following reports support CSV:
  1. Patient Registrations
  2. Cash Payments
  3. Patient Invoices
  4. Chart of Accounts
  5. Voucher Registry