(Feat.) Implement standard infrastructure for assigning filenames
This commit introduces a design for allowing individual report
modules to assign a file name that will be used when the file
is downloaded using the recently introduced bh-download component.

It also uses the HTML5 `download` attribute on an <a> tag and removes
infrastructure required for saving files to a blob Object.