refactor(cash): use CashForm and nicer HTML
This commit completely re-writes the bindings in cash payment HTML to
use the CashForm instead of the controller for data.  The CashForm
service is imported, initialized, and bound to the view.  All
calculations are now done in the CashForm service.

It also refactors the Modal.confirm() to use a better message in English
and to have more standard HTML.

Finally, the cash POS receipt now states "MOTIF:" along with any notes
the user put in if necessary.  The text is larger, the indentation has
been removed, and it is made bold to better facilitate printing on a
thermal printer.