feat(voucher): Patient Invoice Debt Transfer (#186)
* feat(voucher): Patient Invoice Debt Transfer

This commit implements a payment of a patient's debts via the Simple
Voucher module.  The debts are written off towards an account - in the
present use case, against an account belonging to a debtor group.

* fix(vouchers): implement clear() on form

This commit cleans up the form simple voucher form's worst hacks and
replaces them with _slightly_ cleaner code.  It also implements a
clear() button.

Closes #191.

* fix(currency): improve translation/rendering

This commit improves the bhCurrencyInput's translation and rendering by
making sure the currency is always rendered in uppercase and using the
more efficient translate directive instead of the translate filter.

* refactor(vouchers): SimpleVoucher use VoucherForm

This commit refactors the Simple Voucher module to use the VoucherForm.
The VoucherForm brings with it numerous advantages:
 1. Better error detection and handling
 2. Built-in caching.
 3. Cleaner separation of view/model

This makes the SimpleVoucher module much easier to extend and operate
in the future.

* fix(vouchers): fix e2e test

This commit fixes a case-sensitive issue in the simple vouchers form.

* fix(voucher): fix general bugs

This commit fixes the simple vouchers to have more maintainable code,
closes some tags, and makes sure that the appropriate values are
triggered when the scan barcode modal closes.