fix(trial_balance) use POST for large params
This commit ensures that a `POST` request is used for all of the trial
balance methods. `GET` requests with paremeters passed in the query
resulted in server errors from attempting to process very long query
strings.

This commit also udpates the trial balance controller to bring it more
in line with the style of the current repository. Favouring tested
library methods over custom utilities and using more up to date ES6
syntax.