feat(bhAccountSelect): implement component select
This commit implements the bhAccountSelect Component.  The component
takes in an initial accountId and a callback for hooking up the
account_id once the value is set.  It also sets the `$bhValue` so that
`util.filterFormElements()` behaves as expected.

The component looks like this:
```html
<bh-account-select
  account-id="Controller.model.account_id"
  name="account_id"
  on-select-callback="Controller.onSelectAccount(account)">
</bh-account-select>
```

There is an optional `label` field that sets the translation label, and
`name` property to determine the `$bhValue` name.