fix(debtor groups): make sure list is ordered
This commit makes sure the debtor group list is ordered when it returns
from the server.  It also adds hooks for UUID-based selection of the DOM
elements for zoning in on debtor groups.  The previous tests were flaky
because they used `.first()` methods on groups of elements which could
be randomly ordered.

A special debtor group has been created for the delete test.