refactor(db): remove entity_type from gl+pj
This commit removes the column `entity_type` from the general ledger and
the posting journal.  In order to do this, many of the tests for trial
balance failure are no longer relevant.  I have either removed or
skipped them until we can write stricter tests.  Since the posting
journal is being actively worked on, this seems appropriate.

Partially addresses #584.