deps(eslint): rm jshint; install eslint
This commit removes our dependency on jshint and replaces it with eslint
as a linter.  Eslint is much stricter and will reduce the work of a code
reviewer in commenting on style.  The current codebase will need to be
migrated significantly and rulesets customized for this to be an
effective tool.

Closes #223.