refactor(links) Link parent modules in headercrumb
This commit adds anchor tags to the headercrumb*(tm)* module title for
the following modules:
* Patient editing -> Patient Registry
* Patient records -> Patient Registry
* Debtor Group Management Update -> Debtor Group Managment Lists

It also updates the `<a>` style for the headercrumb to differentiate a
link. This style can be refined in the future.

This commit also fixes a number of links that were broken with
dependency updates.