feat(receipts) Add service line to invoice receipt
This commit adds the service that the invoice was created for to both
the invoice POS and standard receipt. On the standard receipt it
replaces the duplicated amount field, this is so that there is one point
of truth for the total amount on the receipt, it can be undone if
required.