fix(registry) Fix refersh issue on invoices/cash
This commit ensures that when an invoice or cash payment is reversed the
registry still filters appropriately.

It also resolves a bug where filters could not be removed when passed in
through a parameter link. This was resolved by standardising passing
parameters by loading them in a startup method and not complicating the
load/ search logic.

Cash payments previously used many different techniques for filtering
results. It has been standardised and now caches filters.