feature(vouchers): implement FilterParsing
This commit implements the FilterParser on the voucher registry and also
updates the integration test to behave as expected.  This includes
searching for the voucher by the full reference and not simply the
reference number.

It also removes the getVouchers() code and replaces it with
lookupVoucher().  This is much cleaner and results in a nicer looking
receipt.

Finally, it refactors the voucher report to be a bit nicer to view/read.